This should solve the issue of the failing tests (old test removed/replaced, and the verdi computer test moved to the right file)
